+++ test/unit/connection/test-debug-parser-error.js
@@ -2,6 +2,7 @@ var assert     = require('assert');
 var common     = require('../../common');
 var connection = common.createConnection({debug: true, port: common.fakeServerPort});
 
+var tid    = 0;
 var server = common.createFakeServer();
 
 server.listen(common.fakeServerPort, function (err) {
@@ -21,10 +22,10 @@ server.listen(common.fakeServerPort, function (err) {
     assert.deepEqual(messages, [
       '<-- HandshakeInitializationPacket',
       '--> ClientAuthenticationPacket',
-      '<-- OkPacket',
-      '--> ComQueryPacket',
-      '<-- ResultSetHeaderPacket',
-      '<-- FieldPacket'
+      '<-- (1) OkPacket',
+      '--> (1) ComQueryPacket',
+      '<-- (1) ResultSetHeaderPacket',
+      '<-- (1) FieldPacket'
     ]);
 
     connection.destroy();
@@ -33,7 +34,7 @@ server.listen(common.fakeServerPort, function (err) {
 });
 
 server.on('connection', function(conn) {
-  conn.handshake();
+  conn.handshake({ threadId: ++tid });
   conn.on('query', function(packet) {
     switch (packet.sql) {
       case 'SELECT value FROM stuff':