+++ test/unit/connection/test-debug.js
@@ -2,6 +2,7 @@ var assert     = require('assert');
 var common     = require('../../common');
 var connection = common.createConnection({debug: true, port: common.fakeServerPort});
 
+var tid    = 0;
 var server = common.createFakeServer();
 
 server.listen(common.fakeServerPort, function (err) {
@@ -21,12 +22,16 @@ server.listen(common.fakeServerPort, function (err) {
     assert.deepEqual(messages, [
       '<-- HandshakeInitializationPacket',
       '--> ClientAuthenticationPacket',
-      '<-- OkPacket',
-      '--> ComPingPacket',
-      '<-- OkPacket'
+      '<-- (1) OkPacket',
+      '--> (1) ComPingPacket',
+      '<-- (1) OkPacket'
     ]);
 
     connection.destroy();
     server.destroy();
   });
+});
+
+server.on('connection', function (conn) {
+  conn.handshake({ threadId: ++tid });
 });